refactor(monitor): move fetching of logs to job-manager (#423)
